Split tests into separate files, add resiliency
