+++ cypress.json
@@ -1,3 +1,3 @@
 {
-  "defaultCommandTimeout": 10000
+  "defaultCommandTimeout": 15000
 }
